# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SEPLAG API Rest",
        "description": "API Rest criada para o processo seletivo da SEPLAG 2025.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/lotacoes": {
            "get": {
                "tags": [
                    "Lotações"
                ],
                "summary": "Listar todas as lotações",
                "description": "Retorna uma lista paginada das lotações cadastradas",
                "operationId": "b286c099696cf17bdfe6e238c407cefd",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página (padrão: 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Quantidade de registros por página (padrão: 10, máximo: 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de lotações",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lotacao"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "currentPage": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "totalPages": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "perPage": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Token inválido ou expirado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lotações"
                ],
                "summary": "Cadastrar nova lotação",
                "description": "Cria um novo registro de lotação",
                "operationId": "272647d30130932cd4cf1a504f30662b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pes_id",
                                    "unid_id",
                                    "lot_data_lotacao"
                                ],
                                "properties": {
                                    "pes_id": {
                                        "type": "integer",
                                        "example": 23
                                    },
                                    "unid_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "lot_data_lotacao": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-15"
                                    },
                                    "lot_data_remocao": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-03-01"
                                    },
                                    "lot_portaria": {
                                        "type": "string",
                                        "example": "Portaria Nº 123/2025"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lotação cadastrada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lotacao"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Token inválido ou expirado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Lotacao": {
                "title": "Lotacao Model",
                "description": "Lotacao model data",
                "properties": {
                    "lot_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "pes_id": {
                        "type": "integer",
                        "example": 23
                    },
                    "unid_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "lot_data_lotacao": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-01-15"
                    },
                    "lot_data_remocao": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-03-01"
                    },
                    "lot_portaria": {
                        "type": "string",
                        "example": "Portaria Nº 123/2025"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}